fix(lint): fix clippy warnings
- Remove unnecessary cast in chunk.rs
- Simplify identity operations in canvas.rs (1 * scale -> scale)
- Use if let instead of is_some().unwrap() pattern in multiple files
- Use !is_empty() instead of len() >= 1 in console
- Add #[allow] attributes for too_many_arguments and type_complexity

+++ src/chunk.rs
@@ -1310,7 +1310,7 @@ impl Chunk {
                 // Pack custom data from model_data
                 for (idx, data) in &self.model_data {
                     let offset = idx * 4;
-                    let bytes = (data.custom_data as u32).to_le_bytes();
+                    let bytes = data.custom_data.to_le_bytes();
                     buf[offset..offset + 4].copy_from_slice(&bytes);
                 }
             }
